feat: lock the local routes to a local server
`/local/**` reads and writes the machine the server runs on, so it exists
only when the local entry sets `OPEN_SWE_LOCAL_MODE` — a hosted deployment
of the same build answers 404. That entry binds loopback unconditionally
rather than defaulting to it, so `--host` is gone.
